backend/back_api/pyproject.toml
[project]
name = "back-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Etienne",email = "etienne.micheladj@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.0",
    "core @ file:///C:/Users/etien/FinUni/divers/BTC_ETH_strat/backend/core",
    "fastparquet (>=2025.12.0,<2026.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
]

[tool.poetry]
packages = [{include = "back_api", from = "src"}]

[tool.poetry.dependencies]
core = {develop = true}

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
httpx = ">=0.27"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
